refactor block renderer
